+++ shared/css/generate_packed.sh
@@ -47,6 +47,7 @@ files=(
     "css/static/modal-serverinfo.css"
     "css/static/modal-settings.css"
     "css/static/modal-volume.css"
+    "css/static/overlay-image-preview.css"
 
     "css/static/ts/tab.css"
     "css/static/ts/chat.css"