+++ ChangeLog.md
@@ -1,6 +1,10 @@
 # Changelog:
 * **19.03.20**
     - Using proper icons for the client info
+    - Added an image preview overlay
+    - Added image preview within the chat
+    - Added an image preview to the music bot thumbnails
+    - Added an image preview to client avatars
     
 * **18.03.20**
     - Updated the sound playback mechanism and allowing the native backend to playback sounds via the native interface.